refactor: address CodeRabbit review on batch_trigger
- collectTriggerContext: derive IntentsCount from the same ViewAll snapshot
  used for the boarding/fee aggregates, eliminating the race between Len()
  and ViewAll().
- evalBatchTrigger: log eval failures at warn level (matches the README
  description of failing open).
- batchtrigger.Trigger: keep cel.Program in a named (unexported) field so
  the package's public surface is just Eval/Source.
- batchtrigger.Parse: validate output type via Type.IsExactType instead of
  pointer equality.
- celenv: clarify in the now() comment why the binding is safe (no
  constant-folding optimizer is enabled on BatchTriggerEnv).

+++ pkg/ark-lib/batchtrigger/celenv/functions.go
@@ -8,13 +8,15 @@ import (
 	"github.com/google/cel-go/common/types/ref"
 )
 
-// now() returns the current Unix timestamp in seconds.
+// now() returns the current Unix timestamp in seconds. The binding is
+// evaluated on every call: BatchTriggerEnv does not opt into the constant
+// folding optimizer (cel.OptOptimize), so the value is never folded.
 var nowFunction = cel.Function("now",
 	cel.Overload("now_double",
 		[]*cel.Type{},
 		cel.DoubleType,
 		cel.FunctionBinding(func(_ ...ref.Val) ref.Val {
-			return types.Double(float64(time.Now().Unix()))
+			return types.Double(time.Now().Unix())
 		}),
 	),
 )
